% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_population_error.R
\name{add_population_error}
\alias{add_population_error}
\title{Adds Population Error to Factor Model Data}
\usage{
add_population_error(
  lf_object,
  cfa_method = c("minres", "ml"),
  fit = c("cfi", "rmsea", "rmsr", "raw"),
  misfit = 0,
  error_method = c("cudeck", "yuan")
)
}
\arguments{
\item{lf_object}{Data object from \code{\link[latentFactoR]{simulate_factors}}}

\item{cfa_method}{Character (length = 1).
Method to generate population error.
Defaults to \code{"minres"}.
Available options:

\itemize{

\item{\code{"minres"}}
{Minimum residual}

\item{\code{"ml"}}
{Maximum likelihood}

}}

\item{fit}{Character (length = 1).
Fit index to control population error.
Defaults to \code{"rmsr"}.
Available options:

\itemize{

\item{\code{"cfi"}}
{Comparative fit index}

\item{\code{"rmsea"}}
{Root mean square error of approximation}

\item{\code{"rmsr"}}
{Root mean square residuals}

\item{\code{"raw"}}
{Direct application of error}

}}

\item{misfit}{Numeric (length = 1).
Magnitude of error to add.
Defaults to \code{0}.
General effect sizes range from small (0.10), moderate (0.20), to large (0.30)}

\item{error_method}{Character (length = 1).
Method to control population error.
Description of methods:

\itemize{

\item{\code{"cudeck"}}
{Description coming soon... see Cudeck & Browne, 1992
for more details}

\item{\code{"yuan"}}
{Description coming soon...}

}}
}
\value{
Returns a list containing:

\item{data}{Simulated data from the specified factor model}

\item{population_correlation}{Population correlation matrix with local dependence added}

\item{parameters}{
A list containing the parameters used to generate the data:

\itemize{

\item{\code{factors}}
{Number of factors}

\item{\code{variables}}
{Variables on each factor}

\item{\code{loadings}}
{Loading matrix}

\item{\code{factor_correlations}}
{Correlations between factors}

\item{\code{categories}}
{Categories for each variable}

\item{\code{skew}}
{Skew for each variable}

}

}

\item{population_error}{
A list containing the parameters used to generate population error:

\itemize{

\item{\code{error_correlation}}
{Correlation matrix with population error added
(same as \code{population_correlation})}

\item{\code{fit}}
{Fit measure used to control population error}

\item{\code{delta}}
{Minimum of the objective function corresponding to the misfit value}

\item{\code{misfit}}
{Specified misfit value}

}

}

\item{original_results}{Original \code{lf_object} input into function}
}
\description{
Adds population error to simulated data from \code{\link[latentFactoR]{simulate_factors}}. 
See examples to get started
}
\examples{
# Generate factor data
two_factor <- simulate_factors(
  factors = 2, # factors = 2
  variables = 6, # variables per factor = 6
  loadings = 0.55, # loadings between = 0.45 to 0.65
  cross_loadings = 0.05, # cross-loadings N(0, 0.05)
  correlations = 0.30, # correlation between factors = 0.30
  sample_size = 1000 # number of cases = 1000
)

# Add small population error using Yuan method
two_factor_Yuan <- add_population_error(
  lf_object = two_factor,
  cfa_method = "minres",
  fit = "rmsr", misfit = 0.10,
  error_method = "yuan"
)

# Add small population error using Cudeck method
two_factor_Yuan <- add_population_error(
  lf_object = two_factor,
  cfa_method = "minres",
  fit = "rmsr", misfit = 0.10,
  error_method = "cudeck"
)

}
\references{
Christensen, A. P., Garrido, L. E., & Golino, H. (2022). \cr
Unique variable analysis: A network psychometrics method to detect local dependence. \cr
\emph{PsyArXiv}

Cudeck, R., & Browne, M.W. (1992). \cr
Constructing a covariance matrix that yields a specified minimizer and a specified minimum discrepancy function value. \cr
\emph{Psychometrika}, \emph{57}, 357–369.

Jimenez, M., Abad, F. J., Garcia-Garzon, E., Golino, H., Christensen, A. P., & Garrido, L. E. (2022). \cr
Dimensionality assessment in generalized bi-factor structures: A network psychometrics approach. \cr
\emph{PsyArXiv}
}
\author{
{\code{\link{bifactor}}} authors \cr
Marcos Jimenez,
Francisco J. Abad,
Eduardo Garcia-Garzon,
Vithor R. Franco,
Luis Eduardo Garrido <luisgarrido@pucmm.edu>

{\code{\link{latentFactoR}}} authors \cr
Alexander P. Christensen <alexpaulchristensen@gmail.com>,
Hudson Golino <hfg9s@virginia.edu>,
Luis Eduardo Garrido <luisgarrido@pucmm.edu>
}
