% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_forest_model.R
\docType{data}
\name{factor_forest_model}
\alias{factor_forest_model}
\title{Factor Forest Model from Goretzko and Buhner}
\format{
An \code{\link{xgboost}} random forest model
}
\usage{
data(factor_forest_model)
}
\description{
Pre-trained random forest model to predict the number of 
factors in the data
}
\details{
Obtained from: https://osf.io/mvrau/
}
\examples{
data("factor_forest_model")

}
\references{
Goretzko, D., & Bühner, M. (2022).
Factor retention using machine learning with ordinal data.
\emph{Applied Psychological Measurement}, 01466216221089345.

Goretzko, D., & Buhner, M. (2020).
One model to rule them all? Using machine learning
algorithms to determine the number of factors
in exploratory factor analysis.
\emph{Psychological Methods}, \emph{25}(6), 776-786.
}
\keyword{datasets}
