% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentFactoR.R
\docType{package}
\name{latentFactoR-package}
\alias{latentFactoR}
\alias{latentFactoR-package}
\title{latentFactoR--package}
\description{
Generates data based on latent factor models.
Data can be continuous, polytomous, dichotomous, or mixed.
Skew, cross-loadings, and population error can be added.
All parameters can be manipulated. Data categorization is based on Garrido, Abad, and Ponsoda (2011).
}
\references{
Garrido, L. E., Abad, F. J., & Ponsoda, V. (2011). \cr
Performance of Velicer’s minimum average partial factor retention method with categorical variables. \cr
Educational and Psychological Measurement, 71(3), 551-570.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>,
Maria Dolores Nieto Canaveras <mnietoca@nebrija.es>,
Hudson Golino <hfg9s@virginia.edu>,
Luis Eduardo Garrido <luisgarrido@pucmm.edu>
}
