% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentFactoR.R
\docType{package}
\name{latentFactoR-package}
\alias{latentFactoR}
\alias{latentFactoR-package}
\title{latentFactoR--package}
\description{
Generates data based on latent factor models.
Data can be continuous, polytomous, dichotomous, or mixed.
Skew, cross-loadings, and population error can be added.
All parameters can be manipulated. Data categorization is based on Garrido, Abad, and Ponsoda (2011).
}
\references{
Christensen, A. P., Garrido, L. E., & Golino, H. (2022). \cr
Unique variable analysis: A network psychometrics method to detect local dependence. \cr
\emph{PsyArXiv}

Garrido, L. E., Abad, F. J., & Ponsoda, V. (2011). \cr
Performance of Velicer’s minimum average partial factor retention method with categorical variables. \cr
\emph{Educational and Psychological Measurement}, \emph{71}(3), 551-570.

Golino, H., Shi, D., Christensen, A. P., Garrido, L. E., Nieto, M. D., Sadana, R., ... & Martinez-Molina, A. (2020).
Investigating the performance of exploratory graph analysis and traditional techniques to identify the number of latent factors: A simulation and tutorial.
\emph{Psychological Methods}, \emph{25}(3), 292-320.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>,
Maria Dolores Nieto Canaveras <mnietoca@nebrija.es>,
Hudson Golino <hfg9s@virginia.edu>,
Luis Eduardo Garrido <luisgarrido@pucmm.edu>
}
