% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_tables.R
\docType{data}
\name{skew_tables}
\alias{skew_tables}
\title{Skew Tables}
\format{
A list (length = 5)
}
\usage{
data(skew_tables)
}
\description{
Tables for skew based on the number of categories (2, 3, 4, 5, or 6) in the data
}
\examples{
data("skew_tables")

}
\keyword{datasets}
